# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REDPR документация АПИ",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/favorite/add": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Добавить предмет в избранные",
                "description": "Добавить в избранные",
                "operationId": "additem",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token, ItemID"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "FKOhXAr6Xhx2e6fMdaKZbTOCxCBwLuJDO3j8fYjRoDG9XoAYKQUSPzayU4BM"
                                    },
                                    "ItemID": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товары"
                    }
                }
            }
        },
        "/api/favorite/view": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Посмотреть избранные",
                "description": "Посмотреть избранные",
                "operationId": "view",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "FKOhXAr6Xhx2e6fMdaKZbTOCxCBwLuJDO3j8fYjRoDG9XoAYKQUSPzayU4BM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товары"
                    }
                }
            }
        },
        "/api/favorite/delete": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Убрать предмет из избранных",
                "description": "Убрать из избранные",
                "operationId": "deleteitem",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token, ItemID"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "FKOhXAr6Xhx2e6fMdaKZbTOCxCBwLuJDO3j8fYjRoDG9XoAYKQUSPzayU4BM"
                                    },
                                    "ItemID": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товары"
                    }
                }
            }
        },
...
        "/api/items/description": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Описание",
                "operationId": "description",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Категория",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Описание товара"
                    }
                }
            }
        },
...
    }
}